Move build_resolvers into build_runner_core.
