--- build_resolvers/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: ../analysis_options.yaml
