--- build_resolvers/lib/build_resolvers.dart
@@ -1,5 +1,0 @@
-// Copyright (c) 2018, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-export 'src/resolver.dart' show AnalyzerResolvers;
