--- build_resolvers/pubspec.yaml
@@ -1,28 +1,0 @@
-name: build_resolvers
-version: 3.0.4
-description: Resolve Dart code in a Builder
-repository: https://github.com/dart-lang/build/tree/master/build_resolvers
-resolution: workspace
-
-environment:
-  sdk: ^3.7.0
-
-dependencies:
-  analyzer: '>=7.4.0 <9.0.0'
-  async: ^2.5.0
-  build: '4.0.0'
-  build_runner_core: '9.3.2'
-  collection: ^1.17.0
-  package_config: ^2.0.0
-  path: ^1.8.0
-  pool: ^1.5.0
-  pub_semver: ^2.0.0
-
-dev_dependencies:
-  build_test: ^3.0.0
-  dart_flutter_team_lints: ^3.1.0
-  logging: ^1.0.0
-  test: ^1.16.0
-
-topics:
- - build-runner
