+++ build_runner/pubspec.yaml
@@ -1,5 +1,5 @@
 name: build_runner
-version: 2.7.2
+version: 2.7.3-wip
 description: A build system for Dart code generation and modular compilation.
 repository: https://github.com/dart-lang/build/tree/master/build_runner
 resolution: workspace
@@ -18,7 +18,7 @@ dependencies:
   build: '4.0.0'
   build_config: ">=1.2.0 <1.3.0"
   build_daemon: ^4.0.0
-  build_runner_core: '9.3.2'
+  build_runner_core: '9.4.0-wip'
   built_collection: ^5.1.1
   code_builder: ^4.2.0
   crypto: ^3.0.0
