+++ build_runner_core/lib/src/resolver/analysis_driver.dart
