+++ build_runner_core/lib/src/resolver/internal.dart
