+++ build_runner_core/lib/src/resolver/resolver.dart
@@ -19,12 +19,13 @@ import 'package:analyzer/src/clients/build_resolvers/build_resolvers.dart';
 import 'package:async/async.dart';
 import 'package:build/build.dart';
 import 'package:build/experiments.dart';
-import 'package:build_runner_core/build_runner_core.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as p;
 import 'package:pool/pool.dart';
 
+import '../logging/build_log.dart';
+import '../logging/timed_activities.dart';
 import 'analysis_driver.dart';
 import 'analysis_driver_filesystem.dart';
 import 'analysis_driver_model.dart';
