+++ build_runner_core/lib/src/resolver/shared_resource_pool.dart
