+++ build_runner_core/test/generate/build_step_impl_test.dart
@@ -10,11 +10,11 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:build/build.dart';
-import 'package:build_resolvers/build_resolvers.dart';
 import 'package:build_runner_core/src/generate/build_step_impl.dart';
 import 'package:build_runner_core/src/generate/run_builder.dart';
 import 'package:build_runner_core/src/generate/single_step_reader_writer.dart';
 import 'package:build_runner_core/src/logging/build_log.dart';
+import 'package:build_runner_core/src/resolver/resolver.dart';
 import 'package:build_test/build_test.dart';
 import 'package:package_config/package_config.dart';
 import 'package:test/test.dart';
