+++ build_test/pubspec.yaml
@@ -1,6 +1,6 @@
 name: build_test
 description: Utilities for writing unit tests of Builders.
-version: 3.3.4
+version: 3.3.5-wip
 repository: https://github.com/dart-lang/build/tree/master/build_test
 resolution: workspace
 
@@ -10,9 +10,8 @@ environment:
 dependencies:
   build: '4.0.0'
   build_config: ^1.0.0
-  build_resolvers: '3.0.4'
-  build_runner: '2.7.2'
-  build_runner_core: '9.3.2'
+  build_runner: '2.7.3-wip'
+  build_runner_core: '9.4.0-wip'
   built_collection: ^5.1.1
   crypto: ^3.0.0
   glob: ^2.0.0
