+++ pubspec.yaml
@@ -14,7 +14,6 @@ workspace:
 - build_config
 - build_daemon
 - build_modules
-- build_resolvers
 - build_runner
 - build_runner_core
 - build_test
